# public/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "iTaxCix API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/v1"
        }
    ],
    "paths": {
...
        "/admin/user-role/create": {
            "post": {
                "tags": [
                    "UserRole"
                ],
                "summary": "Asignar rol a usuario",
                "description": "Asigna un rol a un usuario.",
                "operationId": "createUserRole",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRoleCreateRequestDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Asignación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Creado"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/UserRoleResponseDTO"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Errores de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Petición inválida"
                                        },
                                        "error": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Campos requeridos faltantes"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No autorizado"
                                        },
                                        "error": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Token inválido o expirado"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/user-role/update": {
            "put": {
                "tags": [
                    "UserRole"
                ],
                "summary": "Actualizar asignación de rol a usuario",
                "description": "Actualiza la asignación de un rol a un usuario.",
                "operationId": "updateUserRole",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRoleUpdateRequestDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualización exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "OK"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/UserRoleResponseDTO"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/user-role/delete": {
            "delete": {
                "tags": [
                    "UserRole"
                ],
                "summary": "Eliminar asignación de rol a usuario",
                "description": "Elimina la asignación de un rol a un usuario.",
                "operationId": "deleteUserRole",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRoleDeleteRequestDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Eliminación exitosa (sin contenido)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/user-role/list": {
            "get": {
                "tags": [
                    "UserRole"
                ],
                "summary": "Listar asignaciones de roles a usuarios",
                "description": "Lista todas las asignaciones de roles a usuarios.",
                "operationId": "listUserRoles",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Cantidad de elementos por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "OK"
                                        },
                                        "data": {
                                            "properties": {
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/UserRoleResponseDTO"
                                                    }
                                                },
                                                "meta": {
                                                    "properties": {
                                                        "total": {
                                                            "type": "integer",
                                                            "example": 50
                                                        },
                                                        "perPage": {
                                                            "type": "integer",
                                                            "example": 10
                                                        },
                                                        "currentPage": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "lastPage": {
                                                            "type": "integer",
                                                            "example": 5
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UserRoleCreateRequestDTO": {
                "description": "Datos para crear asignación de usuario a rol",
                "properties": {
                    "userId": {
                        "description": "ID del usuario",
                        "type": "integer",
                        "example": 10
                    },
                    "roleId": {
                        "description": "ID del rol",
                        "type": "integer",
                        "example": 2
                    },
                    "active": {
                        "description": "Si la asignación está activa",
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "UserRoleDeleteRequestDTO": {
                "description": "Datos para eliminar asignación de usuario a rol",
                "properties": {
                    "id": {
                        "description": "ID de la asignación",
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "UserRoleResponseDTO": {
                "description": "Datos de respuesta de asignación de usuario a rol",
                "properties": {
                    "id": {
                        "description": "ID de la asignación",
                        "type": "integer",
                        "example": 1
                    },
                    "userId": {
                        "description": "ID del usuario",
                        "type": "integer",
                        "example": 10
                    },
                    "roleId": {
                        "description": "ID del rol",
                        "type": "integer",
                        "example": 2
                    },
                    "active": {
                        "description": "Si la asignación está activa",
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "UserRoleUpdateRequestDTO": {
                "description": "Datos para actualizar asignación de usuario a rol",
                "properties": {
                    "id": {
                        "description": "ID de la asignación",
                        "type": "integer",
                        "example": 1
                    },
                    "userId": {
                        "description": "ID del usuario",
                        "type": "integer",
                        "example": 10
                    },
                    "roleId": {
                        "description": "ID del rol",
                        "type": "integer",
                        "example": 2
                    },
                    "active": {
                        "description": "Si la asignación está activa",
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}